refactor(agent): extract Instagram inspiration handler
